+++ spec/field_spec.rb
@@ -2,6 +2,9 @@ require 'spec_helper'
 
 describe Contentful::Field do
   let(:field) { vcr('field') { create_client.content_type('cat').fields.first } }
+  let(:linkField) { vcr('linkField') {
+    create_client.content_type('cat').fields.select { |f| f.id == 'image' }.first
+  } }
 
   describe 'Properties' do
     it 'has a #properties getter returning a hash with symbol keys' do
@@ -31,6 +34,16 @@ describe Contentful::Field do
 
     it 'has #localized' do
       expect(field.required).to be_truthy
+    end
+  end
+
+  describe 'Link field properties' do
+    it 'has #type' do
+      expect(linkField.type).to eq 'Link'
+    end
+
+    it 'has #linkType' do
+      expect(linkField.link_type).to eq 'Asset'
     end
   end
 end